requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
Django==1.8.6
django-celery==3.1.17
djangorestframework==3.3.1
hiredis==0.2.0
kombu==3.0.29
psycopg2==2.6.1
pytz==2015.7
redis==2.10.5
requests==2.8.1
wheel==0.24.0
